+++ apps/client/redux/services/auth-service.ts
@@ -3,8 +3,8 @@ import axios from "axios";
 import { UserInfo, UserRolesEnum } from "@/types";
 import { TManagerSignupSchema, TSignupSchema } from "@/lib/validation-schemas";
 import toast from "react-hot-toast";
+import { API_URL } from "../store";
 
-const API_URL = "http://127.0.0.1:4000/api";
 
 interface LoginInput {
   email: string;
@@ -95,43 +95,5 @@ export const registerManager = createAsyncThunk<void, ManagerRegistrationData>(
       }
     }
     toast.success("Registered successfully");
-  }
-);
-
-interface UpdateUserData {
-  id: string;
-  name: string;
-  email: string;
-  newPassword: string;
-  phoneNumber: string;
-  profilePic: File;
-}
-
-export const updateUserProfile = createAsyncThunk<UpdateUserData, UserInfo>(
-  "auth/updateUserProfile",
-  async ({id, name, email, newPassword, phoneNumber, profilePic}, { rejectWithValue }) => {
-    try {
-      const { data } = await axios.patch<UserInfo>(`${API_URL}/user/${id}`, {
-        id,
-        name,
-        email,
-        newPassword,
-        phoneNumber,
-        profilePic
-      });
-      
-
-      // Handle the response accordingly
-      toast.success("Profile updated successfully!");
-      return data;
-    } catch (error: any) {
-      if (error.response && error.response.data.message) {
-        toast.error(error.response.data.message);
-        return rejectWithValue(error.response.data.message);
-      } else {
-        toast.error("An error occurred while updating the profile.");
-        return rejectWithValue(error.message);
-      }
-    }
   }
 );
