+++ apps/client/redux/store.ts
@@ -2,6 +2,8 @@ import { configureStore } from "@reduxjs/toolkit";
 import authReducer from "./features/auth-slice";
 import { TypedUseSelectorHook, useSelector } from "react-redux";
 
+export const API_URL = "http://127.0.0.1:4000/api";
+
 export const store = configureStore({
   reducer: {
     authReducer,
